separate pdf generation components from editor, reduce chances of propogation
