+++ apps/tenant-management-webapp/src/app/store/pdf/reducers.ts
@@ -27,7 +27,6 @@ const defaultState: PdfState = {
 };
 
 export default function (state: PdfState = defaultState, action: PdfActionTypes): PdfState {
-  console.log(JSON.stringify(action.type) + ' <action.type');
   switch (action.type) {
     case FETCH_PDF_TEMPLATES_SUCCESS_ACTION:
       return {
@@ -102,7 +101,6 @@ export default function (state: PdfState = defaultState, action: PdfActionTypes)
           jobs = [action.payload].concat(jobs);
         }
       }
-      console.log(JSON.stringify(jobs) + '<---jobs---------');
 
       return {
         ...state,
